[feat] 플레이어 장비 스탯 연동, CSV 캐릭터 스탯 타입 수정, 이펙트 에셋 추가
EquipmentSlotManager 주석해제와 이벤트 추가

+++ Assets/2. Scripts/CsvData/Data/Character_StatsData.cs
@@ -16,7 +16,7 @@ public class Character_StatsData : ICSVLoad, ITableKey
     public float Character_Def_M { get; set; }
     public float Character_Hp_Regen { get; set; }
     public float Character_Agi { get; set; }
-    public float Character_Level_Exp_N { get; set; }
+    public double Character_Level_Exp_N { get; set; }
 
     int ITableKey.Id => Character_Id;
     string ITableKey.Key => Character_Id.ToString();
@@ -48,6 +48,6 @@ public class Character_StatsData : ICSVLoad, ITableKey
         // 11: Character_Agi (float)
         if (values.Length > 11 && float.TryParse(values[11], out float v11)) Character_Agi = v11;
         // 12: Character_Level_Exp_N (float)
-        if (values.Length > 12 && float.TryParse(values[12], out float v12)) Character_Level_Exp_N = v12;
+        if (values.Length > 12 && double.TryParse(values[12], out double v12)) Character_Level_Exp_N = v12;
     }
 }
